Conflicts:
	src/main/java/hudson/plugins/ec2/SlaveTemplate.java
	src/main/resources/hudson/plugins/ec2/SlaveTemplate/config.jelly
	src/test/java/hudson/plugins/ec2/SlaveTemplateTest.java
	src/test/java/hudson/plugins/ec2/TemplateLabelsTest.java

+++ src/main/java/hudson/plugins/ec2/EC2Slave.java
@@ -8,6 +8,7 @@ import hudson.model.Hudson;
 import hudson.model.Slave;
 import hudson.plugins.ec2.ssh.EC2UnixLauncher;
 import hudson.slaves.NodeProperty;
+import hudson.util.ListBoxModel;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -15,11 +16,18 @@ import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import javax.servlet.ServletException;
+
+import org.apache.commons.lang.StringUtils;
 import org.kohsuke.stapler.DataBoundConstructor;
+import org.kohsuke.stapler.QueryParameter;
 
 import com.amazonaws.AmazonClientException;
 import com.amazonaws.services.ec2.AmazonEC2;
+import com.amazonaws.services.ec2.model.AvailabilityZone;
+import com.amazonaws.services.ec2.model.DescribeAvailabilityZonesResult;
 import com.amazonaws.services.ec2.model.InstanceType;
+import com.amazonaws.services.ec2.model.StopInstancesRequest;
 import com.amazonaws.services.ec2.model.TerminateInstancesRequest;
 
 /**
@@ -35,31 +43,35 @@ public final class EC2Slave extends Slave {
     public final String remoteAdmin; // e.g. 'ubuntu'
     public final String rootCommandPrefix; // e.g. 'sudo'
     public final String jvmopts; //e.g. -Xmx1g
+    public final boolean stopOnTerminate;
 
     /**
      * For data read from old Hudson, this is 0, so we use that to indicate 22.
      */
     private final int sshPort;
 
-    public EC2Slave(String instanceId, String description, String remoteFS, int sshPort, int numExecutors, String labelString, String initScript, String remoteAdmin, String rootCommandPrefix, String jvmopts) throws FormException, IOException {
-        this(instanceId, description, remoteFS, sshPort, numExecutors, Mode.NORMAL, labelString, initScript, Collections.<NodeProperty<?>>emptyList(), remoteAdmin, rootCommandPrefix, jvmopts);
+    public static final String TEST_ZONE = "testZone";
+    
+    public EC2Slave(String instanceId, String description, String remoteFS, int sshPort, int numExecutors, String labelString, String initScript, String remoteAdmin, String rootCommandPrefix, String jvmopts, boolean stopOnTerminate) throws FormException, IOException {
+        this(instanceId, description, remoteFS, sshPort, numExecutors, Mode.NORMAL, labelString, initScript, Collections.<NodeProperty<?>>emptyList(), remoteAdmin, rootCommandPrefix, jvmopts, stopOnTerminate);
     }
 
     @DataBoundConstructor
-    public EC2Slave(String instanceId, String description, String remoteFS, int sshPort, int numExecutors, Mode mode, String labelString, String initScript, List<? extends NodeProperty<?>> nodeProperties, String remoteAdmin, String rootCommandPrefix, String jvmopts) throws FormException, IOException {
+    public EC2Slave(String instanceId, String description, String remoteFS, int sshPort, int numExecutors, Mode mode, String labelString, String initScript, List<? extends NodeProperty<?>> nodeProperties, String remoteAdmin, String rootCommandPrefix, String jvmopts, boolean stopOnTerminate) throws FormException, IOException {
         super(instanceId, description, remoteFS, numExecutors, mode, labelString, new EC2UnixLauncher(), new EC2RetentionStrategy(), nodeProperties);
         this.initScript  = initScript;
         this.remoteAdmin = remoteAdmin;
         this.rootCommandPrefix = rootCommandPrefix;
         this.jvmopts = jvmopts;
         this.sshPort = sshPort;
+        this.stopOnTerminate = stopOnTerminate;
     }
 
     /**
      * Constructor for debugging.
      */
     public EC2Slave(String instanceId) throws FormException, IOException {
-        this(instanceId,"debug","/tmp/hudson", 22, 1, Mode.NORMAL, "debug", "", Collections.<NodeProperty<?>>emptyList(), null, null, null);
+        this(instanceId,"debug", "/tmp/hudson", 22, 1, Mode.NORMAL, "debug", "", Collections.<NodeProperty<?>>emptyList(), null, null, null, false);
     }
 
     /**
@@ -100,9 +112,15 @@ public final class EC2Slave extends Slave {
     public void terminate() {
         try {
             AmazonEC2 ec2 = EC2Cloud.get().connect();
-            TerminateInstancesRequest request = new TerminateInstancesRequest(Collections.singletonList(getInstanceId()));
-            ec2.terminateInstances(request);
-            LOGGER.info("Terminated EC2 instance: "+getInstanceId());
+            if (stopOnTerminate) {
+            	StopInstancesRequest request = new StopInstancesRequest(Collections.singletonList(getInstanceId()));
+            	ec2.stopInstances(request);
+                LOGGER.info("Terminated EC2 instance (stopped): "+getInstanceId());
+            } else {
+            	TerminateInstancesRequest request = new TerminateInstancesRequest(Collections.singletonList(getInstanceId()));
+            	ec2.terminateInstances(request);
+                LOGGER.info("Terminated EC2 instance (terminated): "+getInstanceId());
+            }
             Hudson.getInstance().removeNode(this);
         } catch (AmazonClientException e) {
             LOGGER.log(Level.WARNING,"Failed to terminate EC2 instance: "+getInstanceId(),e);
@@ -131,6 +149,33 @@ public final class EC2Slave extends Slave {
         return sshPort!=0 ? sshPort : 22;
     }
 
+    public boolean getStopOnTerminate() {
+        return stopOnTerminate;
+    }
+
+	public static ListBoxModel fillZoneItems(String accessId,
+			String secretKey, String region) throws IOException,
+			ServletException {
+		ListBoxModel model = new ListBoxModel();
+		if (AmazonEC2Cloud.testMode) {
+			model.add(TEST_ZONE);
+			return model;
+		}
+			
+		if (!StringUtils.isEmpty(accessId) && !StringUtils.isEmpty(secretKey) && !StringUtils.isEmpty(region)) {
+			AmazonEC2 client = AmazonEC2Cloud.connect(accessId, secretKey, AmazonEC2Cloud.getEc2EndpointUrl(region));
+			DescribeAvailabilityZonesResult zones = client.describeAvailabilityZones();
+			List<AvailabilityZone> zoneList = zones.getAvailabilityZones();
+			model.add("<not specified>", "");
+			for (AvailabilityZone z : zoneList) {
+				model.add(z.getZoneName(), z.getZoneName());
+			}
+		}
+		return model;
+	}
+
+    
+    
     @Extension
     public static final class DescriptorImpl extends SlaveDescriptor {
         @Override
@@ -142,6 +187,12 @@ public final class EC2Slave extends Slave {
         public boolean isInstantiable() {
             return false;
         }
+
+        public ListBoxModel doFillZoneItems(@QueryParameter String accessId,
+        		@QueryParameter String secretKey, @QueryParameter String region) throws IOException,
+    			ServletException {
+        	return fillZoneItems(accessId, secretKey, region);
+    	}
     }
 
     private static final Logger LOGGER = Logger.getLogger(EC2Slave.class.getName());
